Bugfix: Import for current file creates new files
This bug occurs when using the "Import properties for active file"
action, where the expected behaviour is to import data into the
currently open file, even if the "wikidata entry id" is not defined.

This bug is a huge issue for me, as I have a hierachical folder
structure for my notes. For example, notes on games should go into
`/media/games`, notes on locations are placed in `/places`, and misc
entries are placed in `/concepts`.